# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Battle Rap API
  version: 1.0.0
  description: >
    REST API for managing online rap battles, tournaments, submissions, judging,
    and administration workflows.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/v1/applications:
    post:
      tags: [Artist]
      summary: Submit participation application
      operationId: submitApplication
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationRequest'
      responses:
        '200':
          description: Application submitted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/submissions/{id}:
    patch:
      tags: [Artist]
      summary: Update submission draft
      operationId: updateSubmission
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmissionPatchRequest'
      responses:
        '200':
          description: Submission updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /api/v1/mod/applications:
    get:
      tags: [Moderator]
      summary: List participation applications
      operationId: listApplications
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Applications pending review.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ApplicationListItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/mod/applications/{id}:
    get:
      tags: [Moderator]
      summary: Get application details
      operationId: getApplicationDetail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Application with applicant context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationApplicationDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/mod/applications/{id}/approve:
    post:
      tags: [Moderator]
      summary: Approve application
      operationId: approveApplication
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Application approved.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/mod/applications/{id}/reject:
    post:
      tags: [Moderator]
      summary: Reject application
      operationId: rejectApplication
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  minLength: 1
              required:
                - reason
      responses:
        '204':
          description: Application rejected.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/mod/profile-changes:
    get:
      tags: [Moderator]
      summary: List profile change requests
      operationId: listProfileChanges
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Profile change requests.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProfileChangeListItem'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/mod/profile-changes/{id}:
    get:
      tags: [Moderator]
      summary: Get profile change request
      operationId: getProfileChangeDetail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Profile change request details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationProfileChangeDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/mod/profile-changes/{id}/approve:
    post:
      tags: [Moderator]
      summary: Approve profile change request
      operationId: approveProfileChange
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Profile change approved.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/mod/profile-changes/{id}/reject:
    post:
      tags: [Moderator]
      summary: Reject profile change request
      operationId: rejectProfileChange
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  minLength: 1
              required:
                - reason
      responses:
        '204':
          description: Profile change rejected.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/mod/submissions:
    get:
      tags: [Moderator]
      summary: List submissions pending moderation
      operationId: listSubmissionsForModeration
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: status
          schema:
            type: string
            enum: [submitted, approved]
        - in: query
          name: round_id
          schema:
            type: string
            format: uuid
            nullable: true
        - in: query
          name: search
          schema:
            type: string
            description: Search by display name or full name.
      responses:
        '200':
          description: Submissions awaiting moderation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationSubmissionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/mod/submissions/{id}:
    get:
      tags: [Moderator]
      summary: Get submission moderation detail
      operationId: getSubmissionForModeration
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Submission details for moderation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationSubmission'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/ErrorResponse'
  /api/v1/mod/submissions/{id}/publish:
    post:
      tags: [Moderator]
      summary: Publish submission
      operationId: publishSubmission
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Submission published.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token issued by the API. Required for authenticated endpoints.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: >
        Refresh token cookie issued during login/registration.
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Authentication required or token invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Insufficient permissions.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        detail:
          type: string
        status:
          type: integer
        field_errors:
          type: object
          additionalProperties:
            type: string
      example:
        type: https://httpstatuses.com/422
        title: Payload validation failed.
        detail: Payload validation failed.
        status: 422
        field_errors:
          email: Invalid email.
...
    ModerationApplicationDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
        email:
          type: string
          format: email
        round_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
        city:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        vk_id:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        beat_author:
          type: string
          nullable: true
        audio_id:
          type: string
          format: uuid
          nullable: true
        lyrics:
          type: string
          nullable: true
        profile_city:
          type: string
          nullable: true
        profile_full_name:
          type: string
          nullable: true
        moderator_id:
          type: string
          format: uuid
          nullable: true
        reject_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        reviewed_at:
          type: string
          format: date-time
          nullable: true
    ModerationProfileChangeDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
        email:
          type: string
          format: email
        changes:
          type: object
          additionalProperties: true
        status:
          type: string
        moderator_id:
          type: string
          format: uuid
          nullable: true
        reject_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        reviewed_at:
          type: string
          format: date-time
          nullable: true
    ModerationSubmission:
      type: object
      required:
        - id
        - status
        - round
        - artist
        - audio
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
        submitted_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
        lyrics:
          type: string
          nullable: true
        round:
          type: object
          required:
            - id
            - number
            - kind
            - tournament_id
            - tournament_title
          properties:
            id:
              type: string
              format: uuid
            number:
              type: integer
            kind:
              type: string
            tournament_id:
              type: string
              format: uuid
            tournament_title:
              type: string
        artist:
          type: object
          required:
            - id
            - display_name
            - email
          properties:
            id:
              type: string
              format: uuid
            display_name:
              type: string
            email:
              type: string
              format: email
        audio:
          type: object
          required:
            - id
            - mime
            - status
          properties:
            id:
              type: string
              format: uuid
            mime:
              type: string
            status:
              type: string
            url:
              type: string
              format: uri
              nullable: true
    ModerationSubmissionListResponse:
      type: object
      required:
        - data
        - page
        - limit
        - total
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ModerationSubmission'
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    ApplicationRequest:
      type: object
      properties:
        round_id:
          type: string
          format: uuid
        city:
          type: string
        age:
          type: integer
          minimum: 12
          maximum: 120
        vk_id:
          type: string
        full_name:
          type: string
        beat_author:
          type: string
        audio_id:
          type: string
          format: uuid
        lyrics:
          type: string
    ApplicationStatusResponse:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          description: Application status.
...
    ApplicationListItem:
      type: object
      required:
        - id
        - user_id
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        round_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
        city:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        vk_id:
          type: string
          nullable: true
        full_name:
          type: string
          nullable: true
        audio_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
...
    ProfileChangeListItem:
      type: object
      required:
        - id
        - user_id
        - changes
        - status
        - created_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        changes:
          type: object
          additionalProperties: true
        status:
          type: string
        created_at:
          type: string
          format: date-time
...
    SubmissionPatchRequest:
      type: object
      properties:
        audio_id:
          type: string
          format: uuid
        lyrics:
          type: string
          nullable: true
    SubmissionStatus:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          description: Submission status.
...